backend/pyproject.toml
[tool.poetry]
name = "saasboilerplate"
version = "0.1.1-alpha.0"
description = "This is the backend for the Software-as-a-Service boilerplate app"
authors = ["Lars Schellhas <lars.schellhas@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
django-filter = "^21.1"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-countries = "^7.2.1"
django-cors-headers = "^3.10.0"
django-username-email = "^2.4.2"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.1"
django-extensions = "^3.1.3"
django-money = "^2.1"
drf-app-generators = "^0.1.5"
django-drf-filepond = "^0.4.0"
dj-stripe = "^2.5.1"
django-environ = "^0.8.1"
stripe = "^2.63.0"
gunicorn = "^20.1.0"
pre-commit = "^2.16.0"
djangorestframework-jwt = "^1.11.0"
cryptography = "^36.0.1"
python-jose = "^3.3.0"
cffi = "^1.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.11.1"
pylint-django = "^2.4.4"
debugpy = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
